requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
aiohttp==3.9.1
asyncio-mqtt==0.16.1
pandas==2.1.4
numpy==1.25.2
python-binance==1.0.19
ccxt==4.1.77
ta==0.10.2
python-dotenv==1.0.0
sqlite3
pydantic==2.5.0
python-multipart==0.0.6
cryptography==41.0.8
# ML and AI libraries
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
# Twelve Data API client
twelvedata==0.2.0
# Additional data processing
yfinance==0.2.28
# Model serialization
joblib==1.3.2
# Telegram bot for alerts
python-telegram-bot==20.7
# Redis for caching
redis==5.0.1
# Additional technical analysis
talib-binary==0.4.26